+++ src/main/java/com/sparta/travelconquestbe/api/coupon/controller/CouponController.java
@@ -1,23 +1,14 @@
 package com.sparta.travelconquestbe.api.coupon.controller;
 
-import com.sparta.travelconquestbe.api.coupon.dto.request.CouponCreateRequest;
-import com.sparta.travelconquestbe.api.coupon.dto.respones.CouponCreateResponse;
 import com.sparta.travelconquestbe.api.coupon.dto.respones.CouponSearchResponse;
 import com.sparta.travelconquestbe.api.coupon.service.CouponService;
-import com.sparta.travelconquestbe.common.annotation.AuthUser;
-import com.sparta.travelconquestbe.common.auth.AuthUserInfo;
-import jakarta.validation.Valid;
 import jakarta.validation.constraints.Positive;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.validation.annotation.Validated;
-import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
@@ -37,23 +28,5 @@ public class CouponController {
   ) {
     Page<CouponSearchResponse> response = couponService.searchAllCoupons(page, limit);
     return ResponseEntity.status(HttpStatus.OK).body(response);
-  }
-
-  @PostMapping("/admins/coupons")
-  public ResponseEntity<CouponCreateResponse> createCoupon(
-      @Valid @RequestBody CouponCreateRequest request,
-      @AuthUser AuthUserInfo user
-  ) {
-    couponService.createCoupon(request, user);
-    return ResponseEntity.noContent().build();
-  }
-
-  @DeleteMapping("/admins/coupons/{id}")
-  public ResponseEntity<String> deleteCoupon(
-      @PathVariable Long id,
-      @AuthUser AuthUserInfo user
-  ) {
-    couponService.deleteCounpon(id, user);
-    return ResponseEntity.status(HttpStatus.NO_CONTENT).body("해당 쿠폰이 삭제되었습니다.");
   }
 }
